requirements.txt
boto3~=1.18.1
dateparser~=1.1.0
netCDF4~=1.5.7
pystac~=1.1.0
python-dateutil~=2.8.2
shapely~=1.7.1
# Keep these up to date with conda-environment.yml:
GDAL~=3.4.2
rasterio~=1.2.6
